--- src/components/columnList/ColumnList.module.scss
@@ -1,221 +1,0 @@
-@import '@/styles/mixin.scss';
-
-.column-list {
-  display: flex;
-  overflow-x: scroll;
-  height: calc(100vh - 6rem);
-  > li {
-    padding-left: 2rem;
-    padding-right: 2rem;
-    padding-bottom: 2rem;
-    background-color: var(--color-gray10);
-    &:not(:last-of-type) {
-      border-right: 2px solid var(--color-gray20);
-    }
-  }
-  @include tablet {
-    flex-direction: column;
-    > li {
-      &:not(:last-of-type) {
-        border-right: 0;
-        border-bottom: 2px solid var(--color-gray20);
-      }
-    }
-  }
-  @include mobile {
-    > li {
-      padding-left: 1.2rem;
-      padding-right: 1.2rem;
-      padding-bottom: 1.2rem;
-    }
-  }
-}
-
-.column-item {
-  flex-basis: 35.4rem;
-  flex-shrink: 0;
-}
-
-.column-title-container {
-  display: grid;
-  grid-template-columns: auto auto auto 1fr;
-  align-items: center;
-  height: 6.8rem;
-  @include mobile {
-    height: 5.4rem;
-  }
-  .dot {
-    width: 0.8rem;
-    height: 0.8rem;
-    margin-right: 0.8rem;
-    border-radius: 50%;
-    background-color: var(--color-violet);
-  }
-  .title {
-    margin-right: 1.2rem;
-    color: var(--color-black_33);
-    font-size: 1.8rem;
-    font-weight: 700;
-    line-height: normal;
-    @include mobile {
-      font-size: 1.6rem;
-    }
-  }
-  .number {
-    width: 2rem;
-    height: 2rem;
-    border-radius: 4px;
-    background-color: var(--color-gray20);
-    font-size: 1.2rem;
-    font-weight: 500;
-    line-height: 2rem;
-    color: var(--color-gray50);
-    text-align: center;
-  }
-  .button-setting-wrapper {
-    justify-self: flex-end;
-    padding: 0;
-    font-size: 0;
-    @include mobile {
-      > img {
-        width: 2.2rem;
-        height: 2.2rem;
-      }
-    }
-  }
-}
-
-.task-list {
-  display: flex;
-  flex-direction: column;
-  gap: 1.6rem;
-  @include mobile {
-    gap: 1rem;
-  }
-}
-
-.task-card-container {
-  display: grid;
-  gap: 1rem 0;
-  align-items: center;
-  padding: 2rem;
-  border: 1px solid var(--color-gray30);
-  border-radius: 6px;
-  background-color: var(--color-white);
-  cursor: pointer;
-  .task-img {
-    width: 100%;
-    height: 16rem;
-    border-radius: 6px;
-    object-fit: cover;
-  }
-  .task-title {
-    color: var(--color-black_33);
-    font-size: 1.6rem;
-    font-weight: 500;
-    line-height: normal;
-  }
-  .tag-list {
-    display: flex;
-    gap: 1rem;
-    .tag {
-      height: 2.2rem;
-      padding: 0.4rem 0.6rem;
-      border-radius: 4px;
-      text-align: center;
-      font-size: 1.2rem;
-      font-weight: 400;
-      line-height: normal;
-    }
-  }
-  .info {
-    display: grid;
-    grid-template-columns: auto auto 1fr;
-    align-items: center;
-    gap: 0.6rem;
-    .date {
-      color: var(--color-gray50);
-      font-size: 1.2rem;
-      font-weight: 500;
-      line-height: normal;
-    }
-    .icon-user {
-      justify-self: end;
-    }
-  }
-  @include tablet {
-    gap: 1rem 1.6rem;
-    grid-template-areas:
-      'img title title'
-      'img tag info';
-    grid-template-columns: auto auto 1fr;
-
-    .task-img {
-      grid-area: img;
-      width: 9rem;
-      margin-right: 0.4rem;
-      height: 5.3rem;
-      object-fit: cover;
-      border-radius: 4px;
-    }
-    .task-title {
-      grid-area: title;
-    }
-    .tag-list {
-      grid-area: tag;
-    }
-    .info {
-      grid-area: info;
-    }
-  }
-  @include mobile {
-    grid-template-columns: 1fr;
-    grid-template-areas:
-      'img'
-      'title'
-      'tag'
-      'info';
-    gap: 0.6rem 0;
-    .task-img {
-      width: 100%;
-      height: 12rem;
-    }
-    .task-title {
-      font-size: 1.4rem;
-    }
-    .tag {
-      font-size: 1.2rem;
-    }
-    .info {
-      gap: 0.4rem;
-      .icon-calender {
-        width: 1.4rem;
-        height: 1.4rem;
-      }
-      .date {
-        font-size: 1rem;
-      }
-      .icon-user {
-        width: 2.2rem;
-        height: 2.2rem;
-      }
-    }
-  }
-}
-
-.add-task-button-container {
-  width: 100%;
-  padding: 0.9rem 0;
-  border: 1px solid var(--color-gray30);
-  border-radius: 6px;
-  background: var(--color-white);
-  text-align: center;
-  font-size: 0;
-  @include mobile {
-    padding: 0.6rem 0;
-    > img {
-      width: 2rem;
-      height: 2rem;
-    }
-  }
-}
