+++ src/components/columns/addTaskButton/AddTaskButton.module.scss
@@ -1,0 +1,18 @@
+@import '@/styles/mixin.scss';
+
+.add-task-button-container {
+  width: 100%;
+  padding: 0.9rem 0;
+  border: 1px solid var(--color-gray30);
+  border-radius: 6px;
+  background: var(--color-white);
+  text-align: center;
+  font-size: 0;
+  @include mobile {
+    padding: 0.6rem 0;
+    > img {
+      width: 2rem;
+      height: 2rem;
+    }
+  }
+}
