+++ src/components/columns/columnItem/ColumnItem.module.scss
@@ -1,0 +1,4 @@
+.column-item {
+  flex-basis: 35.4rem;
+  flex-shrink: 0;
+}
