+++ src/components/columns/columnTitle/ColumnTitle.module.scss
@@ -1,0 +1,50 @@
+@import '@/styles/mixin.scss';
+
+.column-title-container {
+  display: grid;
+  grid-template-columns: auto auto auto 1fr;
+  align-items: center;
+  height: 6.8rem;
+  @include mobile {
+    height: 5.4rem;
+  }
+  .dot {
+    width: 0.8rem;
+    height: 0.8rem;
+    margin-right: 0.8rem;
+    border-radius: 50%;
+    background-color: var(--color-violet);
+  }
+  .title {
+    margin-right: 1.2rem;
+    color: var(--color-black_33);
+    font-size: 1.8rem;
+    font-weight: 700;
+    line-height: normal;
+    @include mobile {
+      font-size: 1.6rem;
+    }
+  }
+  .number {
+    width: 2rem;
+    height: 2rem;
+    border-radius: 4px;
+    background-color: var(--color-gray20);
+    font-size: 1.2rem;
+    font-weight: 500;
+    line-height: 2rem;
+    color: var(--color-gray50);
+    text-align: center;
+  }
+  .button-setting-wrapper {
+    justify-self: flex-end;
+    padding: 0;
+    font-size: 0;
+    @include mobile {
+      > img {
+        width: 2.2rem;
+        height: 2.2rem;
+      }
+    }
+  }
+}
