+++ src/components/columns/taskCardItem/TaskCardItem.module.scss
@@ -1,0 +1,110 @@
+@import '@/styles/mixin.scss';
+
+.task-card-container {
+  display: grid;
+  gap: 1rem 0;
+  align-items: center;
+  padding: 2rem;
+  border: 1px solid var(--color-gray30);
+  border-radius: 6px;
+  background-color: var(--color-white);
+  cursor: pointer;
+  .task-img {
+    width: 100%;
+    height: 16rem;
+    border-radius: 6px;
+    object-fit: cover;
+  }
+  .task-title {
+    color: var(--color-black_33);
+    font-size: 1.6rem;
+    font-weight: 500;
+    line-height: normal;
+  }
+  .tag-list {
+    display: flex;
+    gap: 1rem;
+    .tag {
+      height: 2.2rem;
+      padding: 0.4rem 0.6rem;
+      border-radius: 4px;
+      text-align: center;
+      font-size: 1.2rem;
+      font-weight: 400;
+      line-height: normal;
+    }
+  }
+  .info {
+    display: grid;
+    grid-template-columns: auto auto 1fr;
+    align-items: center;
+    gap: 0.6rem;
+    .date {
+      color: var(--color-gray50);
+      font-size: 1.2rem;
+      font-weight: 500;
+      line-height: normal;
+    }
+    .icon-user {
+      justify-self: end;
+    }
+  }
+  @include tablet {
+    gap: 1rem 1.6rem;
+    grid-template-areas:
+      'img title title'
+      'img tag info';
+    grid-template-columns: auto auto 1fr;
+
+    .task-img {
+      grid-area: img;
+      width: 9rem;
+      margin-right: 0.4rem;
+      height: 5.3rem;
+      object-fit: cover;
+      border-radius: 4px;
+    }
+    .task-title {
+      grid-area: title;
+    }
+    .tag-list {
+      grid-area: tag;
+    }
+    .info {
+      grid-area: info;
+    }
+  }
+  @include mobile {
+    grid-template-columns: 1fr;
+    grid-template-areas:
+      'img'
+      'title'
+      'tag'
+      'info';
+    gap: 0.6rem 0;
+    .task-img {
+      width: 100%;
+      height: 12rem;
+    }
+    .task-title {
+      font-size: 1.4rem;
+    }
+    .tag {
+      font-size: 1.2rem;
+    }
+    .info {
+      gap: 0.4rem;
+      .icon-calender {
+        width: 1.4rem;
+        height: 1.4rem;
+      }
+      .date {
+        font-size: 1rem;
+      }
+      .icon-user {
+        width: 2.2rem;
+        height: 2.2rem;
+      }
+    }
+  }
+}
