+++ src/pages/dashboard/index.tsx
@@ -1,12 +1,10 @@
-import ColumnList from '@/components/columnList/ColumnList'
+import ColumnList from '@/components/columns/columnList/ColumnList'
 import DashboardLayout from '@/components/ui/layout/DashboardLayout'
 
 export default function DashBoardPage() {
   return (
-    <>
-      <DashboardLayout>
-        <ColumnList />
-      </DashboardLayout>
-    </>
+    <DashboardLayout>
+      <ColumnList />
+    </DashboardLayout>
   )
 }
